Feature: Embedded with online help, so that user may always get the most up-to-date help. Bugfix:  Use long instead of int for volume. This is just a partial fix. We still need to do it for the rest.

+++ src/org/yccheok/jstock/gui/MutableStock.java
@@ -79,7 +79,8 @@ public class MutableStock {
         return stock.getLowPrice();
     }
 
-    public int getVolume() {
+    // TODO: CRITICAL LONG BUG REVISED NEEDED.
+    public long getVolume() {
         return stock.getVolume();
     }
